Move to MySQL 8.0 as the default template generation (#11153) (#11769)
* Move towards MySQL 8.0 as the default template generation (#11153)

* Move towards MySQL 8.0 as the default template generation

This upgrades the remaining things to Ubuntu 20.04 and makes MySQL 8.0
the default we run tests against. We still have tests for MySQL 5.7 but
those are now explicitly opted into.

This should finish up the Ubuntu 20.04 upgrade and also makes things
easier for the future when we need to upgrade again.


* CI: rename shard vtorc_8.0 to vtorc_5.7, change expected test output for 8.0


* CI: increase timeout for 8.0 tests on the actual test step from 30 to 45 mins


* CI: increase timeout to 45 minutes for mysql57 tests too. We really only need this for vtorc, but I've made the change to the template so all tests get it.


* CI: fix vtorc test to work with both 5.7 and 8.0


* CI: missed docker flag in mysql57 template, one more fix to vtorc test


* removing spaces from pb file


* removing spaces in pb file part 2



* changing update statement to alter


* Adding mysql 8.0.25


* fix junit

* fixing failures


* fixing more unit tests


* fix selfhost test


* Fixing selfhosted workflows


* fixing vreplication_v2


* removing cluster 12 and 18 from selfhosted


* manual fixing some workflows


* making vreplication_v2 exactly like upstream


* Moving all the workflow to latest mysql version (vanilla)


* remove port restrictions


* tmp fix for vreplication_v2


* making vreplication_v2 to previous version


* correcting typo


* mvoing vreplication_v2 to selfhost


* change docker image to mysql80


* removing limited source from vrepl


* revert back last commit changes


* test: fix aggregation test


* removing ptrtls


* remove pitrTls tests


* changing source/destination to mysql8.0 for vreplication


* remove unwanted workflows


* adding timeout to stress test


* feeback through code review

+++ go/test/endtoend/vtgate/queries/aggregation/aggregation_test.go
@@ -61,9 +61,9 @@ func TestAggregateTypes(t *testing.T) {
 	// TODO (@frouioui): following assertions produce different results between MySQL and Vitess
 	//  their differences are ignored for now. Fix it.
 	// `ascii(val1)` returns an `INT64` on Vitess, and `INT32` on MySQL
-	utils.AssertMatches(t, mcmp.VtConn, "select ascii(val1) as a, count(*) from aggr_test group by a", `[[INT64(65) INT64(1)] [INT64(69) INT64(1)] [INT64(97) INT64(1)] [INT64(98) INT64(1)] [INT64(99) INT64(2)] [INT64(100) INT64(1)] [INT64(101) INT64(1)]]`)
-	utils.AssertMatches(t, mcmp.VtConn, "select ascii(val1) as a, count(*) from aggr_test group by a order by a", `[[INT64(65) INT64(1)] [INT64(69) INT64(1)] [INT64(97) INT64(1)] [INT64(98) INT64(1)] [INT64(99) INT64(2)] [INT64(100) INT64(1)] [INT64(101) INT64(1)]]`)
-	utils.AssertMatches(t, mcmp.VtConn, "select ascii(val1) as a, count(*) from aggr_test group by a order by 2, a", `[[INT64(65) INT64(1)] [INT64(69) INT64(1)] [INT64(97) INT64(1)] [INT64(98) INT64(1)] [INT64(100) INT64(1)] [INT64(101) INT64(1)] [INT64(99) INT64(2)]]`)
+	utils.AssertMatches(t, mcmp.VtConn, "select ascii(val1) as a, count(*) from aggr_test group by a", `[[INT32(65) INT64(1)] [INT32(69) INT64(1)] [INT32(97) INT64(1)] [INT32(98) INT64(1)] [INT32(99) INT64(2)] [INT32(100) INT64(1)] [INT32(101) INT64(1)]]`)
+	utils.AssertMatches(t, mcmp.VtConn, "select ascii(val1) as a, count(*) from aggr_test group by a order by a", `[[INT32(65) INT64(1)] [INT32(69) INT64(1)] [INT32(97) INT64(1)] [INT32(98) INT64(1)] [INT32(99) INT64(2)] [INT32(100) INT64(1)] [INT32(101) INT64(1)]]`)
+	utils.AssertMatches(t, mcmp.VtConn, "select ascii(val1) as a, count(*) from aggr_test group by a order by 2, a", `[[INT32(65) INT64(1)] [INT32(69) INT64(1)] [INT32(97) INT64(1)] [INT32(98) INT64(1)] [INT32(100) INT64(1)] [INT32(101) INT64(1)] [INT32(99) INT64(2)]]`)
 
 	mcmp.AssertMatches("select val1 as a, count(*) from aggr_test group by a", `[[VARCHAR("a") INT64(2)] [VARCHAR("b") INT64(1)] [VARCHAR("c") INT64(2)] [VARCHAR("d") INT64(1)] [VARCHAR("e") INT64(2)]]`)
 	mcmp.AssertMatches("select val1 as a, count(*) from aggr_test group by a order by a", `[[VARCHAR("a") INT64(2)] [VARCHAR("b") INT64(1)] [VARCHAR("c") INT64(2)] [VARCHAR("d") INT64(1)] [VARCHAR("e") INT64(2)]]`)
